+++ TheGame/Player.cs
@@ -5,7 +5,7 @@ namespace TheGame
         public Vector Location { get; set; }
         public double Direction { get; private set; }
         private double TurnAngle { get; set; } = 0.07;
-        public int Size { get; private set; }
+        public float Size { get; private set; }
         public int Speed { get; private set; }
         public double SpeedFactor { get; }
         public int Health { get; set; } = 10;
